Improve Node
Improve the Node struct with two important things: a handle to a mesh,
and a list of handles to its children nodes. Both of them can be invalid
or empty. Imagine a node with no mesh, used to group other nodes
together, or a node which is acting like a camera or a light.