Implement the renderer
Add implementation for two key methods of the Renderer structure: draw
and present. The draw method is particular as it does not render
immediately, but builds a list of nodes to render together with their
transform matrix. The present method is then going to take advantage of
that list by issuing binding and drawing commands in an optimal order.