Improve Pack
Introduce a list of indices to elements, and a list of free positions
into the list of indices. This creates a level of indirection so that
handles point to indices instead of elements. With this approach we are
able to remove elements from the pack easily while keeping the
contiguous property of the pack.