Faster terrain
Generalize instancing by pre-calculating model matrices and binding an
array which is indexed by the instance ID. With this the grass vertex
shader is a level of magnitude faster. We also draw instances in batches
taking into account that we can only bind a limited number of matrices
per draw call.